requirements.txt
# Core Deep Learning
tensorflow>=2.13.0
keras>=2.13.0

# Web Interfaces
gradio>=4.0.0
streamlit>=1.28.0

# Data Processing & Scientific Computing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Image Processing
pillow>=10.0.0
opencv-python>=4.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: For CUDA support (uncomment if using GPU)
# tensorflow-gpu>=2.13.0
